+++ envision/web/src/helpers/state_unpacker.js
@@ -53,6 +53,7 @@ const Traffic = Object.freeze({
   MISSION_ROUTE_GEOMETRY: 11,
   ACTOR_TYPE: 12,
   VEHICLE_TYPE: 13,
+  INTEREST: 14,
 });
 
 const TrafficSignal = Object.freeze({
@@ -146,6 +147,7 @@ function unpack_traffic(traffic) {
         ),
         actor_type: AGENT_TYPE_MAP[t[Traffic.ACTOR_TYPE]],
         vehicle_type: VEHICLE_TYPE_MAP[t[Traffic.VEHICLE_TYPE]],
+        interest: t[Traffic.INTEREST],
       },
     }))
   );