+++ scenarios/sumo/intersections/1_to_2lane_left_turn_c_agents_1/scenario.py
@@ -23,7 +23,15 @@ from itertools import combinations
 from pathlib import Path
 
 from smarts.sstudio import gen_scenario
-from smarts.sstudio.types import Flow, Mission, Route, Scenario, Traffic, TrafficActor
+from smarts.sstudio.types import (
+    Flow,
+    Mission,
+    Route,
+    Scenario,
+    Traffic,
+    TrafficActor,
+    TrapEntryTactic,
+)
 
 normal = TrafficActor(
     name="car",
@@ -86,7 +94,9 @@ route = Route(begin=("E0", 0, 5), end=("E1", 0, "max"))
 ego_missions = [
     Mission(
         route=route,
-        start_time=4,  # Delayed start, to ensure road has prior traffic.
+        entry_tactic=TrapEntryTactic(
+            start_time=4
+        ),  # Delayed start, to ensure road has prior traffic.
     )
 ]