+++ scenarios/sumo/merge/3lane_agents_1/scenario.py
@@ -23,7 +23,15 @@ from itertools import combinations
 from pathlib import Path
 
 from smarts.sstudio import gen_scenario
-from smarts.sstudio.types import Flow, Mission, Route, Scenario, Traffic, TrafficActor
+from smarts.sstudio.types import (
+    Flow,
+    Mission,
+    Route,
+    Scenario,
+    Traffic,
+    TrafficActor,
+    TrapEntryTactic,
+)
 
 normal = TrafficActor(
     name="car",
@@ -71,12 +79,12 @@ for name, routes in enumerate(route_comb):
         ]
     )
 
-route = Route(begin=("gneE6", 0, 10), end=("gneE4", 2, "max"))
+
 ego_missions = [
     Mission(
-        route=route,
-        start_time=15,  # Delayed start, to ensure road has prior traffic.
-    )
+        Route(begin=("gneE6", 0, 10), end=("gneE4", 2, "max")),
+        entry_tactic=TrapEntryTactic(start_time=15),
+    ),
 ]
 
 gen_scenario(