+++ scenarios/sumo/merge/3lane_agents_2/scenario.py
@@ -23,7 +23,15 @@ from itertools import combinations
 from pathlib import Path
 
 from smarts.sstudio import gen_scenario
-from smarts.sstudio.types import Flow, Mission, Route, Scenario, Traffic, TrafficActor
+from smarts.sstudio.types import (
+    Flow,
+    Mission,
+    Route,
+    Scenario,
+    Traffic,
+    TrafficActor,
+    TrapEntryTactic,
+)
 
 normal = TrafficActor(
     name="car",
@@ -73,8 +81,14 @@ for name, routes in enumerate(route_comb):
 
 
 ego_missions = [
-    Mission(Route(begin=("gneE6", 0, 10), end=("gneE4", 2, "max")), start_time=15),
-    Mission(Route(begin=("gneE3", 0, 10), end=("gneE4", 0, "max")), start_time=15),
+    Mission(
+        Route(begin=("gneE6", 0, 10), end=("gneE4", 2, "max")),
+        entry_tactic=TrapEntryTactic(start_time=15),
+    ),
+    Mission(
+        Route(begin=("gneE3", 0, 10), end=("gneE4", 0, "max")),
+        entry_tactic=TrapEntryTactic(start_time=15),
+    ),
 ]
 
 gen_scenario(