+++ scenarios/sumo/zoo_intersection/scenario.py
@@ -13,6 +13,7 @@ from smarts.sstudio.types import (
     SocialAgentActor,
     Traffic,
     TrafficActor,
+    TrapEntryTactic,
 )
 
 # Traffic Vehicles
@@ -96,7 +97,10 @@ gen_scenario(
             f"s-agent-{social_agent1.name}": (
                 [social_agent1],
                 [
-                    EndlessMission(begin=("edge-south-SN", 0, 10), start_time=0.7),
+                    EndlessMission(
+                        begin=("edge-south-SN", 0, 10),
+                        entry_tactic=TrapEntryTactic(start_time=0.7),
+                    ),
                 ],
             ),
         },